feat: Refactor AddToBagTest and RemoveFromBagTest into a single BagTest class with combined functionality
